feat: accountSetting client introduction
